fixed integer percent

+++ Backend/VotingPortal/App/models.py
@@ -16,7 +16,7 @@ class Candidate(models.Model):
     title=models.TextField()
     name=models.TextField()
     vote = models.IntegerField(default='0')
-    percent=models.IntegerField(default='0')
+    percent=models.FloatField(default='0')
     image=models.ImageField()
 
     def __str__(self):
